chore: file structure, separate subsystem module files + fix: indentation in sd, race condition on remount + feat: gps data logging
