+++ flight_computer/main.py
@@ -1,18 +1,21 @@
 from time import sleep
 import _thread
 
-from modules.micropyGPS import MicropyGPS
-from modules.altimeter.altimeter import Altimeter
-from modules.altimeter.sdcard_manager import SDCardManager
-from modules.board import *
+from flight_computer.libs.micropyGPS import MicropyGPS
+from flight_computer.modules.altimeter import Altimeter
+from flight_computer.modules.sdcard_manager import SDCardManager
+from flight_computer.modules.gps import GPS
+from flight_computer.modules.lora import LORA
 
 # Shared SD card manager
 sd_manager = SDCardManager()
+
 
 # THREAD FUNCTIONS
 def radio_task():
     gps = GPS()
     lora = LORA()
+    gps_file_path = "/sd/gps_data.txt"
 
     while True:
         gps_data: MicropyGPS = gps.get_gps_obj()
@@ -20,9 +23,10 @@ def radio_task():
             f"Time: {gps_data.timestamp}, Latitude: {gps_data.latitude}, Longitude: {gps_data.longitude}, Altitude: {gps_data.altitude}"
         )
         lora.transmit_message(gps_msg)
-        # todo: log msg to SD card
+        sd_manager.log(gps_file_path, gps_msg)  # log msg to SD card
 
         sleep(0.5)
+
 
 def altimeter_task():
     altimeter = Altimeter(sd_manager)
