+++ flight_computer/modules/gps.py
@@ -1,22 +1,9 @@
 from machine import Pin, UART
-from modules.ulora import LoRa
-from modules.micropyGPS import MicropyGPS
+from flight_computer.libs.micropyGPS import MicropyGPS
 
 # BOARD DEFNS
 GPS_TX_PIN = 12
 GPS_RX_PIN = 13
-
-LORA_CHANNEL = 0
-LORA_SCK = 18
-LORA_MOSI = 19
-LORA_MISO = 16
-LORA_CS = 17
-LORA_DIO0 = 20  # indicates if TX/RX done
-LORA_RST = 21  # reset pin
-
-# self defined addresses (must match with ground): : 0-254
-LORA_TX_ADDR = 0
-LORA_RX_ADDR = 1
 
 
 # MODULE CLASSES
@@ -52,20 +39,3 @@ class GPS(object):
     def get_gps_obj(self):
         self._set_gps_obj()
         return self.gps_parser
-
-
-class LORA(object):
-    def __init__(self):
-        self.lora = LoRa(
-            (LORA_CHANNEL, LORA_SCK, LORA_MOSI, LORA_MISO),
-            LORA_DIO0,
-            LORA_TX_ADDR,
-            LORA_CS,
-            LORA_RST,
-        )
-
-    def transmit_message(self, message: str):
-        self.lora.send(message, LORA_RX_ADDR)
-        if not self.lora.wait_packet_sent():
-            # todo: log error in the SD card
-            pass
