+++ flight_computer/test_scripts/altimeter_test.py
@@ -5,21 +5,21 @@ Test script for the Altimeter (BMP280) module.
 """
 
 from time import sleep
-from modules.altimeter.altimeter import Altimeter
-from modules.altimeter.sdcard_manager import SDCardManager
+from flight_computer.modules.altimeter import Altimeter
+from flight_computer.modules.sdcard_manager import SDCardManager
 
 sd_manager = SDCardManager()
 altimeter = Altimeter(sd_manager)
-    
+
 while True:
     temperature, pressure = altimeter._read_data()
-        
+
     if temperature is not None and pressure is not None:
         print(f"Temperature: {temperature:.2f} °C, Pressure: {pressure:.2f} Pa")
     else:
         print("Error reading altimeter sensor data.")
-    
+
     altimeter_data = f"Temperature: {temperature:.2f} °C, Pressure: {pressure:.2f} Pa"
     altimeter.sd_manager.log("../sd/altimeter_data.txt", altimeter_data)
-        
+
     sleep(0.5)
